refactor: 프로젝트 상세 페이지 폴더 이름 변경
