+++ frontend/src/pages/project-detail/DetailPage.tsx
