+++ frontend/src/pages/project-detail/components/Carousel/Carousel.styled.ts
