+++ frontend/src/pages/project-detail/components/Carousel/hooks/useSlide.ts
