+++ frontend/src/pages/project-detail/components/SectionTitle.tsx
