+++ frontend/src/pages/project-detail/components/TechStacksSection/TechStacksSection.tsx
